refactor(Filter): Refactored code
