+++ src/components/views/mine-oppgaver/Filter.tsx
@@ -1,0 +1,71 @@
+import { Autocomplete, Box, Button, TextField, ToggleButton, ToggleButtonGroup } from '@material-ui/core';
+import { makeStyles } from '@material-ui/styles';
+import Typo from 'components/Typo';
+import { useData } from 'context/Data';
+import theme from 'theme';
+import { IProcessTemplate, ITag } from 'utils/types';
+
+const useStyles = makeStyles({
+  gutterBottom: {
+    marginBottom: theme.spacing(2),
+  },
+});
+
+type FilterProps = {
+  choosenTags: ITag[];
+  setChoosenTags: (tag: ITag[]) => void;
+  choosenProcessTemplates: string[];
+  setChoosenProcessTemplates: (element: string[]) => void;
+};
+const Filter = ({ choosenTags, setChoosenTags, choosenProcessTemplates, setChoosenProcessTemplates }: FilterProps) => {
+  const { tags, processTemplates } = useData();
+  const clearFilters = () => {
+    setChoosenTags([]);
+    setChoosenProcessTemplates([]);
+  };
+
+  const handleFormat = (_, newFormats) => {
+    if (newFormats.length === processTemplates.length) {
+      setChoosenProcessTemplates([]);
+    } else {
+      setChoosenProcessTemplates(newFormats);
+    }
+  };
+  const classes = useStyles();
+  return (
+    <Box display='flex' flexDirection='column' maxWidth='400px' minWidth='300px' padding={2}>
+      <Typo gutterBottom variant='h2'>
+        Tags
+      </Typo>
+      <Autocomplete
+        className={classes.gutterBottom}
+        getOptionLabel={(option: ITag) => option.title}
+        multiple
+        noOptionsText='Finner ingen tags'
+        onChange={(_, value: ITag[]) => {
+          setChoosenTags(value);
+        }}
+        options={tags}
+        renderInput={(params) => <TextField {...params} size='small' />}
+        value={choosenTags}
+      />
+      <Typo gutterBottom variant='h2'>
+        Prosess
+      </Typo>
+      <ToggleButtonGroup className={classes.gutterBottom} onChange={handleFormat} value={choosenProcessTemplates}>
+        {processTemplates?.map((processTemplate: IProcessTemplate) => {
+          return (
+            <ToggleButton key={processTemplate.id} value={processTemplate.title}>
+              {processTemplate.title}
+            </ToggleButton>
+          );
+        })}
+      </ToggleButtonGroup>
+      <Button disabled={!choosenProcessTemplates.length && !choosenTags.length} onClick={clearFilters} variant='outlined'>
+        Tøm filtre
+      </Button>
+    </Box>
+  );
+};
+
+export default Filter;
