+++ src/pages/mine-oppgaver.tsx
@@ -1,20 +1,20 @@
 import 'moment/locale/nb';
 
-import { Autocomplete, Box, Button, makeStyles, TextField, ToggleButton, ToggleButtonGroup } from '@material-ui/core';
+import { makeStyles } from '@material-ui/core';
 import SearchFilter from 'components/SearchFilter';
 import Typo from 'components/Typo';
+import Filter from 'components/views/mine-oppgaver/Filter';
 import TimeSection from 'components/views/mine-oppgaver/TimeSection';
-import { useData } from 'context/Data';
 import prisma from 'lib/prisma';
 import moment from 'moment';
 import { GetServerSideProps, InferGetServerSidePropsType } from 'next';
 import Head from 'next/head';
 import { useRouter } from 'next/router';
-import { useEffect, useState } from 'react';
+import { useMemo, useState } from 'react';
 import React from 'react';
 import safeJsonStringify from 'safe-json-stringify';
 import theme from 'theme';
-import { IEmployeeTask, IProcessTemplate, ITag } from 'utils/types';
+import { IEmployeeTask, ITag } from 'utils/types';
 import { filterAndSearchTasks, splitIntoTimeSections } from 'utils/utils';
 
 const useStyles = makeStyles({
@@ -116,67 +116,14 @@ const MyTasks = ({ myTasks }: InferGetServerSidePropsType<typeof getServerSidePr
 
   const timeSections: TimeSectionType[] = splitIntoTimeSections(myTasks);
 
-  const clearFilters = () => {
-    setChoosenTags([]);
-    setChoosenProcessTemplates([]);
-  };
-
-  const [RENAMETHIS, SETRENAMETHIS] = useState<TimeSectionType[]>();
-  useEffect(() => {
-    SETRENAMETHIS([]);
-  }, [router.query]);
-  const { tags, processTemplates } = useData();
   const [choosenProcessTemplates, setChoosenProcessTemplates] = useState<string[]>([]);
   const [choosenTags, setChoosenTags] = useState<ITag[]>([]);
   const [searchString, setSearchString] = useState('');
 
-  useEffect(() => {
-    const noe = filterAndSearchTasks(searchString, { tags: choosenTags, processTemplates: choosenProcessTemplates }, timeSections);
-    SETRENAMETHIS(noe);
-  }, [choosenTags, choosenProcessTemplates, searchString]);
+  const filterResult = useMemo(() => {
+    return filterAndSearchTasks(searchString, { tags: choosenTags, processTemplates: choosenProcessTemplates }, timeSections);
+  }, [searchString, choosenTags, choosenProcessTemplates]);
 
-  const handleFormat = (_, newFormats) => {
-    if (newFormats.length === processTemplates.length) {
-      setChoosenProcessTemplates([]);
-    } else {
-      setChoosenProcessTemplates(newFormats);
-    }
-  };
-
-  const FilterComponent = () => (
-    <Box display='flex' flexDirection='column' maxWidth='400px' minWidth='300px' padding={2}>
-      <Typo gutterBottom variant='h2'>
-        Tags
-      </Typo>
-      <Autocomplete
-        className={classes.gutterBottom}
-        getOptionLabel={(option: ITag) => option.title}
-        multiple
-        noOptionsText='Finner ingen tags'
-        onChange={(_, value: ITag[]) => {
-          setChoosenTags(value);
-        }}
-        options={tags}
-        renderInput={(params) => <TextField {...params} size='small' />}
-        value={choosenTags}
-      />
-      <Typo gutterBottom variant='h2'>
-        Prosess
-      </Typo>
-      <ToggleButtonGroup className={classes.gutterBottom} onChange={handleFormat} value={choosenProcessTemplates}>
-        {processTemplates?.map((processTemplate: IProcessTemplate) => {
-          return (
-            <ToggleButton key={processTemplate.id} value={processTemplate.title}>
-              {processTemplate.title}
-            </ToggleButton>
-          );
-        })}
-      </ToggleButtonGroup>
-      <Button disabled={!choosenProcessTemplates.length && !choosenTags.length} onClick={clearFilters} variant='outlined'>
-        Tøm filtre
-      </Button>
-    </Box>
-  );
   return (
     <>
       <Head>
@@ -191,14 +138,21 @@ const MyTasks = ({ myTasks }: InferGetServerSidePropsType<typeof getServerSidePr
         </div>
         <SearchFilter
           activeFilters={Boolean(choosenTags.length || choosenProcessTemplates.length)}
-          filterComponent={<FilterComponent />}
+          filterComponent={
+            <Filter
+              choosenProcessTemplates={choosenProcessTemplates}
+              choosenTags={choosenTags}
+              setChoosenProcessTemplates={setChoosenProcessTemplates}
+              setChoosenTags={setChoosenTags}
+            />
+          }
           search={setSearchString}
         />
         <div>
           {!timeSections.length ? (
             <Typo>Ingen oppgaver</Typo>
           ) : (
-            (RENAMETHIS?.length ? RENAMETHIS : timeSections).map((section: TimeSectionType, index: number) => {
+            (filterResult?.length ? filterResult : timeSections).map((section: TimeSectionType, index: number) => {
               return <TimeSection first={index === 0} key={index} section={section} />;
             })
           )}
